Extract Emitter and Compiler into focused modules to prevent god objects
Split both classes into concern-based modules:

Emitter: ActiveRecordAttributes, ObjectAttributes, MethodFields, Associations
Compiler: ActiveRecordMethods, ObjectMethods, MethodFields, Associations, Dispatch

Each module is a separate file (~50-80 lines) focused on one concern.
Prepares the structure for adding hash-path variants alongside their
JSON counterparts without bloating a single file.
